+++ src/adapter/graphql/tests/tests/test_gql_dataset_flow_runs.rs
@@ -452,6 +452,7 @@ async fn test_trigger_reset_root_dataset_flow() {
         &create_root_result.dataset_handle.id,
         &root_dataset_blocks[1].0,
         &root_dataset_blocks[0].0,
+        false,
         "RESET",
     );
 
@@ -568,8 +569,11 @@ async fn test_trigger_reset_root_dataset_flow() {
                                         },
                                         "startCondition": null,
                                         "configSnapshot": {
-                                            "newHeadHash": &root_dataset_blocks[1].0,
+                                            "mode": {
+                                                "newHeadHash": &root_dataset_blocks[1].0,
+                                            },
                                             "oldHeadHash": &root_dataset_blocks[0].0,
+                                            "recursive": false
                                         }
                                     }
                                 ],
@@ -609,6 +613,7 @@ async fn test_trigger_reset_root_dataset_flow_with_invalid_head() {
         &create_root_result.dataset_handle.id,
         &new_invalid_head,
         &old_invalid_head,
+        false,
         "RESET",
     );
 
@@ -651,6 +656,7 @@ async fn test_trigger_reset_root_dataset_flow_with_invalid_head() {
         &create_root_result.dataset_handle.id,
         &root_dataset_blocks[0].0,
         &root_dataset_blocks[1].0,
+        false,
         "RESET",
     );
 
@@ -3499,9 +3505,17 @@ impl FlowRunsHarness {
                                                 }
                                                 __typename
                                             }
-                                             ... on FlowConfigurationReset {
-                                                newHeadHash
+                                            ... on FlowConfigurationReset {
+                                                mode {
+                                                    ... on SnapshotConfigurationResetCustom {
+                                                        newHeadHash
+                                                    }
+                                                    ... on SnapshotConfigurationResetToSeedDummy {
+                                                        dummy
+                                                    }
+                                                }
                                                 oldHeadHash
+                                                recursive
                                             }
                                             ... on FlowConfigurationCompactionRule {
                                                 compactionRule {
@@ -3626,7 +3640,7 @@ impl FlowRunsHarness {
                                                             }
                                                         }
                                                     }
-                                                }
+                                               }
                                             }
                                         }
                                     }
@@ -3646,6 +3660,7 @@ impl FlowRunsHarness {
         id: &DatasetID,
         new_head_hash: &Multihash,
         old_head_hash: &Multihash,
+        recursive: bool,
         dataset_flow_type: &str,
     ) -> String {
         indoc!(
@@ -3659,8 +3674,13 @@ impl FlowRunsHarness {
                                     datasetFlowType: "<dataset_flow_type>",
                                     flowRunConfiguration: {
                                         reset: {
-                                            newHeadHash: "<new_head_hash>",
-                                            oldHeadHash: "<old_head_hash>"
+                                            mode: {
+                                                custom: {
+                                                    newHeadHash: "<new_head_hash>"
+                                                }
+                                            },
+                                            oldHeadHash: "<old_head_hash>",
+                                            recursive: <recursive>
                                         }
                                     }
                                 ) {
@@ -3706,6 +3726,7 @@ impl FlowRunsHarness {
         .replace("<dataset_flow_type>", dataset_flow_type)
         .replace("<new_head_hash>", &new_head_hash.to_string())
         .replace("<old_head_hash>", &old_head_hash.to_string())
+        .replace("<recursive>", if recursive { "true" } else { "false" })
     }
 
     fn trigger_flow_with_compaction_config_mutation(