+++ src/domain/task-system/domain/src/entities/logical_plan.rs
@@ -82,7 +82,8 @@ pub struct HardCompactionDataset {
 pub struct ResetDataset {
     pub dataset_id: DatasetID,
     pub new_head_hash: Option<Multihash>,
-    pub old_head_hash: Multihash,
+    pub old_head_hash: Option<Multihash>,
+    pub recursive: bool,
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////